# agent/04_openapi.yaml
openapi: 3.0.3
info:
  title: Tour Booking Platform API (MVP)
  version: 0.1.0
  description: |
    REST API target contract for the Tour Booking Platform MVP.
    Current implementation status and known drift are tracked in `agent/13_api_status_matrix.md`.
    As of 2026-03-18, this file still contains target-only paths for release-gap work such as
    email verification, password reset, announcements, reports, moderation, and richer review summaries.
servers:
  - url: https://api.example.com
security:
  - bearerAuth: []

# =========================
# 1) Auth + Me + Notifications
# Paste under: paths:
# =========================
paths:
...
  /public/announcements:
    get:
      tags: [Announcements]
      security: []
      summary: List public announcements
      parameters:
        - name: organizationId
          in: query
          schema: { $ref: '#/components/schemas/Id64' }
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnnouncementListResponse' }
  /operator/organizations/{orgId}/announcements:
    get:
      tags: [Announcements]
      summary: List operator announcements (ORG_ADMIN/ORG_OWNER)
      parameters:
        - $ref: '#/components/parameters/orgId'
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnnouncementListResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /organizations/{orgId}/announcements:
    post:
      tags: [Announcements]
      summary: Create announcement (ORG_ADMIN/ORG_OWNER)
      parameters:
        - $ref: '#/components/parameters/orgId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AnnouncementCreateRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Announcement' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /announcements/{announcementId}:
    patch:
      tags: [Announcements]
      summary: Update announcement (ORG_ADMIN/ORG_OWNER)
      parameters:
        - $ref: '#/components/parameters/announcementId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AnnouncementPatchRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Announcement' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Unprocessable' }

    delete:
      tags: [Announcements]
      summary: Delete announcement (ORG_ADMIN/ORG_OWNER)
      parameters:
        - $ref: '#/components/parameters/announcementId'
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
# =========================
# 5) Bookings + Waitlist + Participants + Notes + Policies
# =========================
...
  /organizations/{orgId}/reports/occurrences:
    get:
      tags: [Reports]
      summary: Occurrence ops report (ORG_ADMIN/ORG_OWNER)
      parameters:
        - $ref: '#/components/parameters/orgId'
        - $ref: '#/components/parameters/dateFromOptional'
        - $ref: '#/components/parameters/dateToOptional'
        - name: tourId
          in: query
          schema: { $ref: '#/components/schemas/Id64' }
        - name: occurrenceId
          in: query
          schema: { $ref: '#/components/schemas/Id64' }
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OccurrenceReportResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /organizations/{orgId}/reports/occurrences/export:
    get:
      tags: [Reports]
      summary: Export occurrence ops report CSV (ORG_ADMIN/ORG_OWNER)
      parameters:
        - $ref: '#/components/parameters/orgId'
        - $ref: '#/components/parameters/dateFromOptional'
        - $ref: '#/components/parameters/dateToOptional'
        - name: tourId
          in: query
          schema: { $ref: '#/components/schemas/Id64' }
        - name: occurrenceId
          in: query
          schema: { $ref: '#/components/schemas/Id64' }
      responses:
        '200':
          description: OK
          content:
            text/csv:
              schema:
                type: string
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /organizations/{orgId}/reports/bookings:
    get:
      tags: [Reports]
      summary: Booking report (ORG_ADMIN/ORG_OWNER)
      parameters:
        - $ref: '#/components/parameters/orgId'
        - $ref: '#/components/parameters/dateFromOptional'
        - $ref: '#/components/parameters/dateToOptional'
        - name: tourId
          in: query
          schema: { $ref: '#/components/schemas/Id64' }
        - name: occurrenceId
          in: query
          schema: { $ref: '#/components/schemas/Id64' }
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookingReportResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /organizations/{orgId}/reports/bookings/export:
    get:
      tags: [Reports]
      summary: Export booking report CSV (ORG_ADMIN/ORG_OWNER)
      parameters:
        - $ref: '#/components/parameters/orgId'
        - $ref: '#/components/parameters/dateFromOptional'
        - $ref: '#/components/parameters/dateToOptional'
        - name: tourId
          in: query
          schema: { $ref: '#/components/schemas/Id64' }
        - name: occurrenceId
          in: query
          schema: { $ref: '#/components/schemas/Id64' }
      responses:
        '200':
          description: OK
          content:
            text/csv:
              schema:
                type: string
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ---------
  # Moderation (Admin-only, optional in MVP)
  # ---------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OccurrenceStatus:
      type: string
      enum: [SCHEDULED, CANCELED, FINISHED]
...
    BookingStatus:
      type: string
      enum:
        - REQUESTED
        - WAITLISTED
        - OFFERED
        - CONFIRMED
        - REJECTED
        - CANCELED
        - EXPIRED
        - COMPLETED
    PaymentStatus:
      type: string
      enum:
        - AUTHORIZED
        - PAID
        - REFUND_PENDING
        - REFUNDED
...
    Announcement:
      type: object
      required: [id, organizationId, title, body, visibility, createdAt, updatedAt]
      properties:
        id: { $ref: '#/components/schemas/Id64' }
        organizationId: { $ref: '#/components/schemas/Id64' }
        title: { type: string, maxLength: 200 }
        body: { type: string, maxLength: 5000 }
        visibility:
          type: string
          enum: [DRAFT, PUBLIC, INTERNAL]
        publishStartsAtUtc: { type: string, format: date-time, nullable: true }
        publishEndsAtUtc: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AnnouncementCreateRequest:
      type: object
      required: [title, body, visibility]
      properties:
        title: { type: string, maxLength: 200 }
        body: { type: string, maxLength: 5000 }
        visibility:
          type: string
          enum: [DRAFT, PUBLIC, INTERNAL]
        publishStartsAtUtc: { type: string, format: date-time, nullable: true }
        publishEndsAtUtc: { type: string, format: date-time, nullable: true }
    AnnouncementPatchRequest:
      type: object
      properties:
        title: { type: string, maxLength: 200 }
        body: { type: string, maxLength: 5000 }
        visibility:
          type: string
          enum: [DRAFT, PUBLIC, INTERNAL]
        publishStartsAtUtc: { type: string, format: date-time, nullable: true }
        publishEndsAtUtc: { type: string, format: date-time, nullable: true }
    AnnouncementListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Announcement' }
        nextCursor: { type: string, nullable: true }

    # -----------------------------
    # Reports
    # -----------------------------
    OccurrenceReportItem:
      type: object
      properties:
        occurrenceId: { $ref: '#/components/schemas/Id64' }
        organizationId: { $ref: '#/components/schemas/Id64' }
        tourId: { $ref: '#/components/schemas/Id64' }
        startsAtUtc: { type: string, format: date-time, nullable: true }
        status: { $ref: '#/components/schemas/OccurrenceStatus' }
        capacity: { type: integer, minimum: 0 }
        confirmedSeats: { type: integer, minimum: 0 }
        waitlistCount: { type: integer, minimum: 0 }
        seatUtilizationPercent: { type: integer, minimum: 0, maximum: 100 }
        attendedCount: { type: integer, minimum: 0 }
        noShowCount: { type: integer, minimum: 0 }
        refundedBookingCount: { type: integer, minimum: 0 }
        refundPendingCount: { type: integer, minimum: 0 }
    OccurrenceReportResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/OccurrenceReportItem' }
        nextCursor: { type: string, nullable: true }
    BookingReportItem:
      type: object
      properties:
        bookingId: { $ref: '#/components/schemas/Id64' }
        occurrenceId: { $ref: '#/components/schemas/Id64' }
        tourId: { $ref: '#/components/schemas/Id64' }
        organizerUserId: { $ref: '#/components/schemas/Id64' }
        partySize: { type: integer, minimum: 1 }
        status: { $ref: '#/components/schemas/BookingStatus' }
        paymentStatus: { $ref: '#/components/schemas/PaymentStatus' }
        refundStatus:
          type: string
          nullable: true
        createdAt: { type: string, format: date-time }
    BookingReportResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/BookingReportItem' }
        nextCursor: { type: string, nullable: true }

    # -----------------------------
    # Policies (Cancellation rules)
    # -----------------------------
...
    Id64:
      type: integer
      format: int64
...
    ErrorCode:
      type: string
      description: Domain/application error code used in `error.code`
      enum:
        - VALIDATION_ERROR
        - REQUIRED_FIELD_MISSING
        - PARTY_SIZE_OUT_OF_RANGE
        - BOOKING_SCOPE_MISMATCH
        - INVITE_WINDOW_CLOSED
        - ATTENDANCE_NOT_ELIGIBLE
        - INVALID_PARTICIPANT_STATUS
        - OFFER_EXPIRY_INVALID
        - INVALID_STATE_TRANSITION
        - BOOKING_TERMINAL_STATE
        - OFFER_NOT_ACTIVE
        - OFFER_EXPIRED
        - OCCURRENCE_ALREADY_CANCELED
        - CAPACITY_EXCEEDED
        - PARTY_SIZE_INCREASE_NOT_ALLOWED
        - WAITLIST_PROMOTION_NOT_POSSIBLE
        - DUPLICATE_REVIEW
        - INQUIRY_ALREADY_EXISTS
        - IDEMPOTENCY_KEY_REUSED_WITH_DIFFERENT_PAYLOAD
        - IDEMPOTENCY_IN_PROGRESS
        - PAYMENT_ALREADY_REFUNDED
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid request
            details:
              type: object
              additionalProperties: true
              nullable: true
...
  responses:
...
    Unauthorized:
      description: Unauthorized (missing/invalid token)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Forbidden (insufficient permission)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Not Found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Unprocessable:
      description: Unprocessable Entity (validation error)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  parameters:
...
    cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string, nullable: true }
      description: Opaque cursor for pagination
    limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      description: Page size
...
    orgId:
      name: orgId
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
    announcementId:
      name: announcementId
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
    dateFromOptional:
      name: dateFrom
      in: query
      required: false
      schema: { type: string, format: date-time }
      description: Inclusive start in UTC
    dateToOptional:
      name: dateTo
      in: query
      required: false
      schema: { type: string, format: date-time }
      description: Exclusive end in UTC
...